refactor: remake folder names
