+++ src/modules/reusable-tests/index.ts
@@ -1,1 +1,1 @@
-export { createReusableTest } from "./utils";
+export { createReusableTest } from "./create-reusable-test";
