+++ src/modules/scenarios/model/report.ts
@@ -1,0 +1,13 @@
+import { TestingTreeSingleton, ConsoleReporter } from "@modules/reporting";
+
+export function report() {
+    const testingTree = TestingTreeSingleton.getInstance().tree;
+
+    testingTree.calculateResults();
+
+    const reporter = new ConsoleReporter();
+
+    reporter.report();
+
+    return testingTree.toObject();
+}
