+++ src/modules/tests/index.ts
@@ -1,6 +1,6 @@
 export { SampleTest } from "./classes/sample-test";
 export { TestsGroup } from "./classes/test-group";
-export { BaseTestFactory } from "./model/base-test-factory";
-export { TestFactory } from "./model/test-factory";
+export { BaseTestFactory } from "./factories/base-test-factory";
+export { TestFactory } from "./factories/test-factory";
 export { TestStatus } from "./types";
 export type { TestInfo, AvailableTestStates } from "./types";
